requirements.txt
# Time Warp II Requirements
# Core dependencies for the Educational Programming IDE

# Runtime dependencies (most included with Python standard library)
# - tkinter (GUI framework)
# - json (configuration management)
# - os, sys (system interface)
# - threading (concurrent execution)
# - subprocess (process management)
# - turtle (graphics engine)
# - pathlib (file system operations)

# Required runtime dependencies
pygame-ce>=2.0.0        # Graphics and multimedia support (community edition)
pygments>=2.15.0        # Syntax highlighting for code editor

# Development dependencies
pytest>=7.0.0          # Testing framework
pytest-cov>=4.0.0      # Coverage plugin used by pyproject pytest addopts
black>=22.0.0           # Code formatting
flake8>=4.0.0          # Linting

# Runtime dependencies for enhanced features
Pillow>=8.0.0        # Image processing (PIL)